refactor: extract browser detection functions to separate file
